Change almost every use of size_t to uint32_t or channel_t to decrease memory transfer.
32 bits should still be more than enough for libmixed's real-time needs,
and for batch processing the data can still just be processed in many rounds.
I thus see no need to use full register precision for almost anything here.

+++ examples/common.h
@@ -88,7 +88,7 @@ int extract_encoding(int encodings){
   return -1;
 }
 
-int load_out_segment(size_t samples, struct out **_out){
+int load_out_segment(uint32_t samples, struct out **_out){
   long out_samplerate = output_samplerate;
   int out_channels = 2;
   int out_encoding = MPG123_ENC_SIGNED_16;
@@ -209,7 +209,7 @@ void free_mp3(struct mp3 *mp3){
   free(mp3);
 }
 
-int load_mp3_segment(char *file, size_t samples, struct mp3 **_mp3){
+int load_mp3_segment(char *file, uint32_t samples, struct mp3 **_mp3){
   long mp3_samplerate = 0;
   int mp3_channels = 0;
   int mp3_encoding = 0;
