+++ src/segment.c
@@ -28,7 +28,7 @@ MIXED_EXPORT int mixed_segment_end(struct mixed_segment *segment){
   return 1;
 }
 
-MIXED_EXPORT int mixed_segment_set_in(size_t field, size_t location, void *value, struct mixed_segment *segment){
+MIXED_EXPORT int mixed_segment_set_in(uint32_t field, uint32_t location, void *value, struct mixed_segment *segment){
   mixed_err(MIXED_NO_ERROR);
   if(segment->set_in)
     return segment->set_in(field, location, value, segment);
@@ -36,7 +36,7 @@ MIXED_EXPORT int mixed_segment_set_in(size_t field, size_t location, void *value
   return 0;
 }
 
-MIXED_EXPORT int mixed_segment_set_out(size_t field, size_t location, void *value, struct mixed_segment *segment){
+MIXED_EXPORT int mixed_segment_set_out(uint32_t field, uint32_t location, void *value, struct mixed_segment *segment){
   mixed_err(MIXED_NO_ERROR);
   if(segment->set_out)
     return segment->set_out(field, location, value, segment);
@@ -44,7 +44,7 @@ MIXED_EXPORT int mixed_segment_set_out(size_t field, size_t location, void *valu
   return 0;
 }
 
-MIXED_EXPORT int mixed_segment_get_in(size_t field, size_t location, void *value, struct mixed_segment *segment){
+MIXED_EXPORT int mixed_segment_get_in(uint32_t field, uint32_t location, void *value, struct mixed_segment *segment){
   mixed_err(MIXED_NO_ERROR);
   if(segment->get_in)
     return segment->get_in(field, location, value, segment);
@@ -52,7 +52,7 @@ MIXED_EXPORT int mixed_segment_get_in(size_t field, size_t location, void *value
   return 0;
 }
 
-MIXED_EXPORT int mixed_segment_get_out(size_t field, size_t location, void *value, struct mixed_segment *segment){
+MIXED_EXPORT int mixed_segment_get_out(uint32_t field, uint32_t location, void *value, struct mixed_segment *segment){
   mixed_err(MIXED_NO_ERROR);
   if(segment->get_out)
     return segment->get_out(field, location, value, segment);
@@ -67,7 +67,7 @@ MIXED_EXPORT int mixed_segment_info(struct mixed_segment_info *info, struct mixe
   return 0;
 }
 
-MIXED_EXPORT int mixed_segment_get(size_t field, void *value, struct mixed_segment *segment){
+MIXED_EXPORT int mixed_segment_get(uint32_t field, void *value, struct mixed_segment *segment){
   mixed_err(MIXED_NO_ERROR);
   if(segment->get)
     return segment->get(field, value, segment);
@@ -75,7 +75,7 @@ MIXED_EXPORT int mixed_segment_get(size_t field, void *value, struct mixed_segme
   return 0;
 }
 
-MIXED_EXPORT int mixed_segment_set(size_t field, void *value, struct mixed_segment *segment){
+MIXED_EXPORT int mixed_segment_set(uint32_t field, void *value, struct mixed_segment *segment){
   mixed_err(MIXED_NO_ERROR);
   if(segment->set)
     return segment->set(field, value, segment);
